requirements.txt
Django==2.1.7
psycopg2==2.7.7
psycopg2-binary==2.7.7
django-widget-tweaks==1.4.3
django-debug-toolbar==1.11
Pillow==5.4.1
djangorestframework==3.9.2